изменен тип данных номера проекта внесены изменения в соответствующие поля обоих дто

+++ src/main/java/ru/trae/backend/dto/project/ProjectDto.java
@@ -15,7 +15,7 @@ import ru.trae.backend.dto.operation.OperationDto;
  */
 public record ProjectDto(
     long id,
-    long number,
+    int number,
     String name,
     LocalDateTime startDate,
     LocalDateTime plannedEndDate,
